# openapi.yaml
openapi: 3.1.0
info:
  title: Status Page Service API
  version: 1.0.0
  description: >
    REST API сервиса страниц статуса (аналог Статусмейт / Statuspal).
    Manual-first: статусы компонентов и инциденты управляются вручную; автоматизация
    (входящие webhook'и от Grafana/Prometheus/PagerDuty и generic) — опциональный слой.
    Все временные метки — ISO 8601 в UTC. Нормативный источник — DESIGN.md (разделы 5–7).
  license:
    name: Proprietary
servers:
  - url: https://api.example.ru/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local dev
...
security:
  - BearerAuth: []
  - ApiToken: []
paths:
...
  /pages/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Pages]
      summary: Получить страницу (управление)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Pages]
      summary: Изменить страницу
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StatusPageUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Pages]
      summary: Удалить страницу
      responses:
        '204': { description: No Content }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Операторский access-токен (JWT) в заголовке `Authorization: Bearer <token>`.
        Получается через /auth/login или /auth/refresh.
    ApiToken:
      type: apiKey
      in: header
      name: Authorization
      description: >
        Токен страницы передаётся как значение заголовка Authorization (без префикса Bearer),
        совместимо с поведением Статусмейт.
    IntegrationHmac:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-подпись тела запроса секретом интеграции.
  parameters:
...
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Не авторизован / неверный токен
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Конфликт (например, домен/slug уже заняты)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    StatusPage:
      type: object
      required: [id, name, slug, visibility]
      properties:
        id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        slug: { type: string }
        timezone: { type: string }
        default_locale: { type: string }
        visibility: { type: string, enum: [public, private] }
        custom_domain: { type: [string, "null"] }
        domain_verified: { type: boolean }
        theme: { type: object, additionalProperties: true }
        logo_url: { type: [string, "null"] }
        favicon_url: { type: [string, "null"] }
        hide_powered_by: { type: boolean }
        redirect_url: { type: [string, "null"] }
        # Адрес отправителя писем (премиум). Сам smtp_config (с паролем) в ответах НЕ отдаётся —
        # вместо него булев флаг smtp_configured.
        from_email: { type: [string, "null"] }
        smtp_configured: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    StatusPageUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        timezone: { type: string }
        default_locale: { type: string }
        visibility: { type: string, enum: [public, private] }
        # Пароль приватной страницы (write-only). Непустая строка — задать/сменить;
        # null — снять пароль. В ответах не возвращается (хранится только хэш, §9).
        password: { type: [string, "null"] }
        # Собственный домен (этап 4.3). Непустая строка — задать (сбрасывает domain_verified
        # до повторной проверки CNAME); null/"" — снять. Уникален среди страниц.
        custom_domain: { type: [string, "null"] }
        theme: { type: object, additionalProperties: true }
        logo_url: { type: [string, "null"] }
        favicon_url: { type: [string, "null"] }
        hide_powered_by: { type: boolean }
        redirect_url: { type: [string, "null"] }
        # Кастомный SMTP страницы (этап 4.5, премиум). Объект {host,port,username,password,from,tls}
        # — задать; null — снять (вернуться на системный SMTP). Write-only (в ответах не отдаётся).
        smtp_config: { type: [object, "null"], additionalProperties: true }
        # Адрес отправителя писем (этап 4.5). Строка — задать; null/"" — снять.
        from_email: { type: [string, "null"] }
...